requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
flask>=2.0.0
mysql-connector-python>=8.0.0
scikit-learn>=1.0.0
joblib>=1.0.0